Split out the main metal file.
